Added Testing and refactored
